# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for storing and download PDFs into a Google Cloud Plattform bucket",
        "title": "Library API",
        "contact": {
            "name": "Leonardo Bispo",
            "email": "leonardobispo.dev@gmail.com"
        },
        "license": {
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/books": {
            "get": {
                "description": "Get books matching passed filters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Books"
                ],
                "summary": "Get Books",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name of the book you want to search",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ID of the book you want to search",
                        "name": "book_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.BookResponseDoc"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creates the book in database and uploads the file into bucket",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Books"
                ],
                "summary": "Create Book",
                "parameters": [
                    {
                        "description": "Required book information",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.BookRequestDoc"
                        }
                    },
                    {
                        "type": "file",
                        "description": "Book File",
                        "name": "book",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.BookResponseDoc"
                        }
                    }
                }
            }
        },
        "/books/download": {
            "post": {
                "description": "Download your books by passing the necessary parameters",
                "tags": [
                    "Cloud"
                ],
                "summary": "Download Book",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the book you want to download",
                        "name": "book_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.BookRequestDoc": {
            "type": "object",
            "properties": {
                "bookInfo": {
                    "$ref": "#/definitions/handler.bookInfo"
                }
            }
        },
        "handler.BookResponseDoc": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "format": "date",
                    "example": "2006-01-02T15:04:05Z"
                },
                "deleted_at": {
                    "type": "string",
                    "format": "date",
                    "example": "2006-01-02T15:04:05Z"
                },
                "description": {
                    "type": "string",
                    "example": "Some description here"
                },
                "gender": {
                    "type": "string",
                    "example": "notice"
                },
                "id": {
                    "type": "string",
                    "example": "36e4379a-be55-462d-ab9e-da0c532623ea"
                },
                "release_date": {
                    "type": "string",
                    "example": "2023"
                },
                "title": {
                    "type": "string",
                    "example": "Test title"
                },
                "updated_at": {
                    "type": "string",
                    "format": "date",
                    "example": "2006-01-02T15:04:05Z"
                },
                "writer": {
                    "type": "string",
                    "example": "Jhonn Doe"
                }
            }
        },
        "handler.bookInfo": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Some description here"
                },
                "gender": {
                    "type": "string",
                    "example": "notice"
                },
                "release_date": {
                    "type": "string",
                    "example": "2023"
                },
                "title": {
                    "type": "string",
                    "example": "Test title"
                },
                "writer": {
                    "type": "string",
                    "example": "Jhonn Doe"
                }
            }
        }
    }
}